Comments and Restructuring
Added comments in a few places

Restructured data retrieval for the different projects. It was previously in several different arrays (image links, github links, etc), but now it is in one big 2D array to help with easier rearranging and possibly better performance.

I also added const keywords on a few variables, such as the github warning text